Лабораторная работа №4

+++ Lab04/lab04/04-01.js
@@ -17,6 +17,15 @@ db.on('POST', (req, res) => {
     console.log('DB POST');
     req.on('data', (data) => {
         let r = JSON.parse(data);
+
+        console.log(r);
+
+        if (r.id === '' || r.name === '' || r.bday === '') {
+            res.writeHead(400, {'Content-Type': 'application/json; charset=utf-8'});
+            res.end(JSON.stringify({error: 'Поля не могут быть пустыми'}));
+            return;
+        }
+
         db.insert(r).then(data => {
             res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'});
             res.end(JSON.stringify(data));
@@ -31,6 +40,13 @@ db.on('PUT', (req, res) => {
     console.log('DB PUT');
     req.on('data', (data) => {
         let r = JSON.parse(data);
+
+        if (r.id === '' || r.name === '' || r.bday === '') {
+            res.writeHead(400, {'Content-Type': 'application/json; charset=utf-8'});
+            res.end(JSON.stringify({error: 'Поля не могут быть пустыми'}));
+            return;
+        }
+
         db.update(r).then(data => {
             res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'});
             res.end(JSON.stringify(data));
@@ -46,6 +62,13 @@ db.on('DELETE', (req, res) => {
     console.log('DB DELETE');
     req.on('data', (data) => {
         let r = JSON.parse(data);
+
+        if (r.id === '' || r.name === '' || r.bday === '') {
+            res.writeHead(400, {'Content-Type': 'application/json; charset=utf-8'});
+            res.end(JSON.stringify({error: 'Поля не могут быть пустыми'}));
+            return;
+        }
+
         db.delete(r.id).then((data) => {
             res.writeHead(200, {'Content-Type': 'application/json; charset=utf-8'});
             res.end(JSON.stringify(data));
